requirements.txt
argparse
beaker
bottle
mwoauth
pymongo
pyyaml
requests
requests-oauthlib
scipy
oursql
